backend/requirements.txt
asgiref==3.8.1
Django==4.2
django-cors-headers==4.5.0
django-filter==23.1
djangorestframework==3.15.2
djangorestframework-simplejwt==4.7.2
drf-yasg==1.21.7
inflection==0.5.1
isort==5.13.2
packaging==24.1
pillow==11.0.0
psycopg2-binary==2.9.9
PyJWT==2.9.0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
sqlparse==0.5.1
tzdata==2024.2
uritemplate==4.1.1
